Move Plan to be an inner class of ResolverRegistry.
