+++ reasoner/resolution/resolver/RuleResolver.java
@@ -163,7 +163,7 @@ public class RuleResolver extends Resolver<RuleResolver> {
         resolvables.addAll(concludablesWithApplicableRules);
         resolvables.addAll(retrievables);
         
-        List<Resolvable> plan = Resolvable.plan(resolvables, conceptMgr, logicMgr);
+        List<Resolvable> plan = ResolverRegistry.plan(resolvables, conceptMgr, logicMgr);
         for (Resolvable planned : plan) {
             Pair<Actor<? extends ResolvableResolver<?>>, Map<Reference.Name, Reference.Name>> concludableUnifierPair = registry.registerResolvable(planned);
             this.plan.add(concludableUnifierPair);
