+++ test/integration/reasoner/BUILD
@@ -42,6 +42,30 @@ host_compatible_java_test(
     ],
 )
 
+host_compatible_java_test(
+    name = "test-resolvable-registry",
+    srcs = ["ResolvableRegistryTest.java"],
+    native_libraries_deps = [
+        "//rocks:rocks",
+        "//:grakn",
+        "//concept:concept",
+        "//pattern:pattern",
+        "//logic:logic",
+        "//reasoner:reasoner",
+    ],
+    test_class = "grakn.core.reasoner.ResolvableRegistryTest",
+    deps = [
+        # Internal dependencies
+        "//test/integration/util",
+        "//common",
+
+        # External dependencies from Grakn Labs
+        "@graknlabs_common//:common",
+        "@graknlabs_graql//java/pattern",
+        "@graknlabs_graql//java:graql",
+    ],
+)
+
 checkstyle_test(
     name = "checkstyle",
     include = glob(["*"]),
